puppet-openstack/HA-CENTOS-UBUNTU/modules/castlight_dashboard/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
# PBR should always appear first
# NOTE(mriedem): Remove this workaround once bug 1375048 is fixed.
testrepository>=0.0.18
pbr!=0.7,<1.0,>=0.6
# Horizon Core Requirements
Babel>=1.3
Django<1.8,>=1.4.2
Pint>=0.5 # BSD
django-compressor<2.0,>=1.4
django-openstack-auth<1.3.0,>=1.2.0
django-pyscss<2.0.0,>=1.0.3 # BSD License (2 clause)
eventlet!=0.17.0,>=0.16.1
httplib2>=0.7.5
iso8601>=0.1.9
kombu>=2.5.0
netaddr>=0.7.12
oslo.concurrency<1.9.0,>=1.8.2 # Apache-2.0
oslo.config<1.10.0,>=1.9.3 # Apache-2.0
oslo.i18n<1.6.0,>=1.5.0 # Apache-2.0
oslo.serialization<1.5.0,>=1.4.0 # Apache-2.0
oslo.utils!=1.4.1,<1.5.0,>=1.4.0 # Apache-2.0
pyScss<1.3,>=1.2.1 # MIT License
python-ceilometerclient<1.2,>=1.1.1
python-cinderclient<1.2.0,>=1.1.0
python-glanceclient<0.18.0,>=0.15.0
python-heatclient<0.5.0,>=0.3.0
python-keystoneclient<1.4.0,>=1.2.0
python-neutronclient<2.5.0,>=2.4.0
python-novaclient<2.24.0,>=2.22.0
python-saharaclient<0.9.0,>=0.8.0
python-swiftclient<2.5.0,>=2.2.0
python-troveclient<1.1.0,>=1.0.7
pytz>=2013.6
PyYAML>=3.1.0
six>=1.9.0
XStatic>=1.0.0 # MIT License
XStatic-Angular>=1.3.7 # MIT License
XStatic-Angular-Bootstrap>=0.11.0.2 # MIT License
XStatic-Angular-lrdragndrop>=1.0.2.2 # MIT License
XStatic-Bootstrap-Datepicker>=1.3.1.0 # Apache 2.0 License
XStatic-Bootstrap-SCSS>=3 # Apache 2.0 License
XStatic-D3>=3.1.6.2 # BSD License (3 clause)
XStatic-Hogan>=2.0.0.2 # Apache 2.0 License
XStatic-Font-Awesome>=4.2.0 # SIL OFL 1.1 License, MIT License
XStatic-Jasmine>=2.1.2.0 # MIT License
XStatic-jQuery>=1.7.2 # MIT License
XStatic-JQuery-Migrate>=1.2.1.1 # MIT License
XStatic-JQuery.quicksearch>=2.0.3.1 # MIT License
XStatic-JQuery.TableSorter>=2.0.5b.0 # MIT License
XStatic-jquery-ui>=1.10.1 # MIT License
XStatic-JSEncrypt>=2.0.0.2 # MIT License
XStatic-Magic-Search>=0.2.0.1 # Apache 2.0 License
XStatic-QUnit>=1.14.0.2 # MIT License
XStatic-Rickshaw>=1.5.0 # BSD License (prior)
XStatic-smart-table>=1.4.5.3 # MIT License
XStatic-Spin>=1.2.5.2 # MIT License
XStatic-term.js>=0.0.4 # MIT License
hurry.filesize>=0.9